web/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.1
django-allauth==0.50.0
django-environ==0.11.2
django-filter==24.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.8
mysqlclient==2.2.4
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.1
style==1.1.0
typing_extensions==4.12.2
update==0.0.1
urllib3==2.2.2
